+++ static/js/server_events_dispatch.js
@@ -263,7 +263,7 @@ export function dispatch_normal_event(event) {
         case "realm_bot":
             if (event.op === "add") {
                 bot_data.add(event.bot);
-            } else if (event.op === "remove") {
+            } else if (event.op === "deactivate") {
                 bot_data.deactivate(event.bot.user_id);
                 event.bot.is_active = false;
             } else if (event.op === "delete") {