+++ zerver/lib/events.py
@@ -630,7 +630,7 @@ def apply_event(
         if event["op"] == "add":
             state["realm_bots"].append(event["bot"])
 
-        if event["op"] == "remove":
+        if event["op"] == "deactivate":
             user_id = event["bot"]["user_id"]
             for bot in state["realm_bots"]:
                 if bot["user_id"] == user_id: